requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via python-debian
charset-normalizer==2.0.6
    # via requests
idna==3.2
    # via requests
python-debian==0.1.42
    # via -r requirements.in
python-gitlab==2.10.1
    # via -r requirements.in
requests==2.26.0
    # via
    #   python-gitlab
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via python-gitlab
urllib3==1.26.6
    # via requests
